requirements.txt
playwright==1.41.0
pytest==7.4.3
pytest-playwright==0.4.3
pytest-xdist==3.5.0
pytest-rerunfailures==12.0
python-dotenv==1.0.0
requests==2.31.0
jsonschema==4.20.0
faker==22.0.0
allure-pytest==2.13.2
